requirements.txt
asgiref==3.2.9
Django==3.0.7
django-cors-headers==3.4.0
djangorestframework==3.11.0
psycopg2==2.8.5
pytz==2020.1
sqlparse==0.3.1
pytest==3.6
pytest-django